complete homepage, optimize header/footer, and migrate CSS to SCSS
